requirements.txt
streamlit>=1.24.0
pymysql>=1.0.3
python-dotenv>=1.0.0
pandas>=2.0.0
